console: Move selector to global scope and refactor
