+++ pkg/webui/account/components/oauth-client-form/messages.js
@@ -55,10 +55,10 @@ export default defineMessages({
   stateDescriptionDesc:
     'You can use this field to save additional information about the state of this OAuth client, e.g. why it has been flagged',
   contactWarning:
-    'Note that if no contact is provided, it will default to the first collaborator of the organization.',
+    'Note that if no contact is provided, it will default to the first collaborator of the client.',
   adminContactDescription:
-    'Administrative contact information for this application. Typically used to indicate who to contact with administrative questions about the application.',
+    'Administrative contact information for this client. Typically used to indicate who to contact with administrative questions about the client.',
   techContactDescription:
-    'Technical contact information for this application. Typically used to indicate who to contact with technical/security questions about the application.',
+    'Technical contact information for this client. Typically used to indicate who to contact with technical/security questions about the client.',
   contactPlaceholder: 'Type to choose a contact',
 })
