+++ pkg/webui/console/containers/organization-form/form.js
@@ -21,23 +21,20 @@ import SubmitBar from '@ttn-lw/components/submit-bar'
 import SubmitButton from '@ttn-lw/components/submit-button'
 import Notification from '@ttn-lw/components/notification'
 
+import CollaboratorSelect from '@ttn-lw/containers/collaborator-select'
+import {
+  decodeContact,
+  encodeContact,
+  organizationSchema,
+  userSchema,
+} from '@ttn-lw/containers/collaborator-select/util'
+
 import Message from '@ttn-lw/lib/components/message'
 
 import Yup from '@ttn-lw/lib/yup'
 import PropTypes from '@ttn-lw/lib/prop-types'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
-import { id as organizationIdRegexp, userId as contactIdRegex } from '@ttn-lw/lib/regexp'
-
-import CollaboratorSelect from '../collaborator-select'
-import { decodeContact, encodeContact } from '../collaborator-select/util'
-
-const organizationSchema = Yup.object().shape({
-  organization_id: Yup.string().matches(contactIdRegex, sharedMessages.validateAlphanum),
-})
-
-const userSchema = Yup.object().shape({
-  user_id: Yup.string().matches(contactIdRegex, sharedMessages.validateAlphanum),
-})
+import { id as organizationIdRegexp } from '@ttn-lw/lib/regexp'
 
 const validationSchema = Yup.object().shape({
   ids: Yup.object().shape({
@@ -72,10 +69,9 @@ const m = defineMessages({
   contactWarning:
     'Note that if no contact is provided, it will default to the first collaborator of the organization.',
   adminContactDescription:
-    'Administrative contact information for this application. Typically used to indicate who to contact with administrative questions about the application.',
+    'Administrative contact information for this organization. Typically used to indicate who to contact with administrative questions about the organization.',
   techContactDescription:
-    'Technical contact information for this application. Typically used to indicate who to contact with technical/security questions about the application.',
-  contactPlaceholder: 'Type to choose a contact',
+    'Technical contact information for this organization. Typically used to indicate who to contact with technical/security questions about the organization.',
 })
 
 const initialValues = {
@@ -127,8 +123,8 @@ const OrganizationForm = props => {
           <Notification small warning content={m.contactWarning} />
           <CollaboratorSelect
             name="administrative_contact"
-            title={'Administrative contact'}
-            placeholder={m.contactPlaceholder}
+            title={sharedMessages.adminContact}
+            placeholder={sharedMessages.contactPlaceholder}
             entity="organization"
             entityId={orgId}
             encode={encodeContact}
@@ -141,8 +137,8 @@ const OrganizationForm = props => {
           />
           <CollaboratorSelect
             name="technical_contact"
-            title={'Technical contact'}
-            placeholder={m.contactPlaceholder}
+            title={sharedMessages.technicalContact}
+            placeholder={sharedMessages.contactPlaceholder}
             entity="organization"
             entityId={orgId}
             encode={encodeContact}
