+++ pkg/webui/console/views/gateway-general-settings/basic-settings-form/validation-schema.js
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { organizationSchema, userSchema } from '@ttn-lw/containers/collaborator-select/util'
+
 import Yup from '@ttn-lw/lib/yup'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
-import { id as gatewayIdRegexp, userId as contactIdRegex } from '@ttn-lw/lib/regexp'
+import { id as gatewayIdRegexp } from '@ttn-lw/lib/regexp'
 
 import {
   attributeValidCheck,
@@ -24,14 +26,6 @@ import {
   attributesCountCheck,
 } from '@console/lib/attributes'
 import { addressWithOptionalScheme as addressWithOptionalSchemeRegexp } from '@console/lib/regexp'
-
-const organizationSchema = Yup.object().shape({
-  organization_id: Yup.string().matches(contactIdRegex, sharedMessages.validateAlphanum),
-})
-
-const userSchema = Yup.object().shape({
-  user_id: Yup.string().matches(contactIdRegex, sharedMessages.validateAlphanum),
-})
 
 const validationSchema = Yup.object().shape({
   ids: Yup.object().shape({
