+++ pkg/webui/containers/collaborator-select/index.js
