+++ pkg/webui/containers/collaborator-select/util.js
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import Yup from '@ttn-lw/lib/yup'
+import sharedMessages from '@ttn-lw/lib/shared-messages'
+import { userId as contactIdRegex } from '@ttn-lw/lib/regexp'
+
 export const composeOption = value => ({
   value:
     'user_ids' in value?.ids
@@ -40,3 +44,11 @@ export const encodeContact = value =>
     : null
 
 export const decodeContact = value => (value ? composeContactOption(value) : null)
+
+export const organizationSchema = Yup.object().shape({
+  organization_id: Yup.string().matches(contactIdRegex, sharedMessages.validateAlphanum),
+})
+
+export const userSchema = Yup.object().shape({
+  user_id: Yup.string().matches(contactIdRegex, sharedMessages.validateAlphanum),
+})
